# distsql: account for aggregate func memory usage