--- a/pkg/sql/sem/builtins/aggregate_builtins_test.go
+++ b/pkg/sql/sem/builtins/aggregate_builtins_test.go
@@ -42,6 +42,7 @@ func testAggregateResultDeepCopy(
 	evalCtx := tree.NewTestingEvalContext(cluster.MakeTestingClusterSettings())
 	defer evalCtx.Stop(context.Background())
 	aggImpl := aggFunc([]types.T{vals[0].ResolvedType()}, evalCtx, nil)
+	defer aggImpl.Close(context.Background())
 	runningDatums := make([]tree.Datum, len(vals))
 	runningStrings := make([]string, len(vals))
 	for i := range vals {
@@ -286,16 +287,19 @@ func runBenchmarkAggregate(
 	params := []types.T{vals[0].ResolvedType()}
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		aggImpl := aggFunc(params, evalCtx, nil)
-		for i := range vals {
-			if err := aggImpl.Add(context.Background(), vals[i]); err != nil {
-				b.Fatal(err)
+		func() {
+			aggImpl := aggFunc(params, evalCtx, nil)
+			defer aggImpl.Close(context.Background())
+			for i := range vals {
+				if err := aggImpl.Add(context.Background(), vals[i]); err != nil {
+					b.Fatal(err)
+				}
 			}
-		}
-		res, err := aggImpl.Result()
-		if err != nil || res == nil {
-			b.Errorf("taking result of aggregate implementation %T failed", aggImpl)
-		}
+			res, err := aggImpl.Result()
+			if err != nil || res == nil {
+				b.Errorf("taking result of aggregate implementation %T failed", aggImpl)
+			}
+		}()
 	}
 }