The size of AggregateFunc implementations was previously not taken into
account by the Aggregator processor's memory monitoring.

Also, some aggregate builtins were not accounting for additional memory
usage during accumulation, particularly those which relied on Decimals, which
can grow arbitrarily large. I added more monitoring where appropriate.

The original issue warns against calling `Grow()` for every row. I'm calling
`ResizeTo()`, which is a no-op when the size hasn't actually changed. That
seems to be the case the vast majority of the time here, so I don't think it's
a bottleneck. There is still some overhead from the method dispatch, which
could be avoided by keeping track of the previous size on the AggregatorFunc
side and only calling `ResizeTo` when it changes. Not sure it's enough of a
performance hit to warrant that though. (We're talking about single digit
nanoseconds per row.) I can add that logic if reviewers feel it's worth it.

```
name                                   old time/op  new time/op  delta
AvgAggregateInt/count=1000-4           92.1µs ± 8%  96.8µs ± 5%     ~     (p=0.151 n=5+5)
AvgAggregateSmallInt/count=1000-4      27.1µs ± 3%  28.1µs ± 3%   +3.92%  (p=0.032 n=5+5)
AvgAggregateFloat/count=1000-4         7.35µs ± 1%  7.55µs ± 2%   +2.86%  (p=0.008 n=5+5)
AvgAggregateDecimal/count=1000-4        152µs ± 3%   162µs ± 2%   +6.15%  (p=0.008 n=5+5)
CountAggregate/count=1000-4            3.50µs ± 1%  3.49µs ± 1%     ~     (p=0.841 n=5+5)
SumIntAggregateInt/count=1000-4        14.8µs ± 2%  14.7µs ± 1%     ~     (p=0.310 n=5+5)
SumAggregateInt/count=1000-4           78.0µs ± 1%  83.4µs ± 1%   +6.92%  (p=0.008 n=5+5)
SumAggregateSmallInt/count=1000-4      21.1µs ± 2%  21.5µs ± 1%   +1.90%  (p=0.032 n=5+5)
SumAggregateFloat/count=1000-4         3.90µs ± 1%  3.99µs ± 1%   +2.32%  (p=0.008 n=5+5)
SumAggregateDecimal/count=1000-4        142µs ± 2%   150µs ± 2%   +5.26%  (p=0.008 n=5+5)
MaxAggregateInt/count=1000-4           8.72µs ± 2%  9.25µs ± 1%   +6.10%  (p=0.008 n=5+5)
MaxAggregateFloat/count=1000-4         8.75µs ± 2%  9.23µs ± 1%   +5.48%  (p=0.008 n=5+5)
MaxAggregateDecimal/count=1000-4       61.3µs ±10%  68.2µs ±27%     ~     (p=0.841 n=5+5)
MinAggregateInt/count=1000-4           8.66µs ± 2%  9.60µs ± 2%  +10.81%  (p=0.008 n=5+5)
MinAggregateFloat/count=1000-4         8.63µs ± 2%  9.38µs ± 2%   +8.67%  (p=0.008 n=5+5)
MinAggregateDecimal/count=1000-4       37.7µs ± 6%  42.8µs ±16%     ~     (p=0.056 n=5+5)
VarianceAggregateInt/count=1000-4      9.18ms ± 2%  9.71ms ± 5%   +5.70%  (p=0.032 n=5+5)
VarianceAggregateFloat/count=1000-4    9.31µs ± 1%  9.26µs ± 2%     ~     (p=0.548 n=5+5)
VarianceAggregateDecimal/count=1000-4  9.22ms ± 1%  9.27ms ± 2%     ~     (p=0.548 n=5+5)
StdDevAggregateInt/count=1000-4        9.20ms ± 1%  9.47ms ± 4%     ~     (p=0.056 n=5+5)
StdDevAggregateFloat/count=1000-4      13.1µs ± 1%  13.8µs ± 3%   +5.77%  (p=0.008 n=5+5)
StdDevAggregateDecimal/count=1000-4    9.33ms ± 2%  9.32ms ± 3%     ~     (p=1.000 n=5+5)
```